% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ctd.R
\name{create_ctd}
\alias{create_ctd}
\title{Create gene expression specificity scores for scRNAseq data}
\usage{
create_ctd(
  matrix = NULL,
  meta_lev_1 = NULL,
  meta_lev_2 = NULL,
  outdir = NULL,
  mhc_genes = NULL,
  group_name = NULL,
  threads = NULL
)
}
\arguments{
\item{matrix}{Raw count gene matrix - needs to be genes x cell and annotation data}

\item{meta_lev_1}{A vector of cell IDs for level 1 cell types}

\item{meta_lev_2}{A vector of cell IDs for level 1 cell types}

\item{outdir}{The output directory to store the ctd object}

\item{mhc_genes}{A set of MHC genes to be removed}

\item{group_name}{An string identifier for the dataset}

\item{threads}{Number of cores to assign to process. Default is NULL,
which will auto detect total cores and estimate number to use.}
}
\value{
A CTD object containing

directory where ctd object was stored
}
\description{
Create gene expression specificity scores for scRNAseq data
}
\examples{
data('seurat_small', 'mhc_genes')
create_ctd(seurat_small@assays$RNA@data,
                     meta_lev_1 = seurat_small$cluster_level_1,
                     meta_lev_2 = seurat_small$cluster_level_2,
                     outdir = 'test',
                     mhc_genes = mhc_genes,
                     group_name = 'brain_study',
                     threads = 1)
}
