% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_enrich_test_files.R
\name{create_enrich_test_files}
\alias{create_enrich_test_files}
\title{Create files for SLDSR and MAGMA gene set enrichment tests from ctd object (Top 10\%)}
\usage{
create_enrich_test_files(
  ctd_path = NULL,
  study_id = NULL,
  gene_coordinates = NULL,
  levels = c(1, 2),
  outdir = NULL,
  sldsr_ext = c(100, 100)
)
}
\arguments{
\item{ctd_path}{A string of file path for ctd object}

\item{study_id}{A unique identifier for the study}

\item{gene_coordinates}{A gene coordinate reference file}

\item{levels}{A vector of levels}

\item{outdir}{A directory to save the enrichment test files to.}

\item{sldsr_ext}{A vector of the upstream and downstream gene / peak boundary
extension window for SLDSR}
}
\value{
Writes a set of tsv MAGMA / SLDSR files to directory of interest
}
\description{
Create files for SLDSR and MAGMA gene set enrichment tests from ctd object (Top 10\%)
}
\examples{
data(gene_coord)
create_enrich_test_files(ctd_path = 'data/ctd_obj.rda',
                                   study_id = 'brain_study',
                                   gene_coordinates = gene_coord,
                                   outdir = 'test/')
}
