% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd_obj}
\alias{ctd_obj}
\title{A CTD object containing cell specificty scores for an snRNAseq dtatset}
\format{
An object of class \code{"list"};
}
\source{
Generated using the EWCE::generate_celltype_data
}
\usage{
data(ctd_obj)
}
\description{
A list
}
\examples{
data(ctd_obj)
}
\references{
Downloaded from
}
\keyword{datasets}
