% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_coord}
\alias{gene_coord}
\title{A Dataframe of genes and their hg19 co-ordinates which have had MHC
overlapping genes removed.}
\format{
An object of class \code{"data.frame"};
}
\source{
\href{https://ctg.cncr.nl/software/MAGMA/aux_files/NCBI37.3.zip}{MAGMA hg19 ref}
}
\usage{
data(gene_coord)
}
\description{
A Dataframe of genes and their genomic co-ordinates taken from a gene reference
file which have had MHC overlapping genes removed. Entrez and HGNC IDs are
included.
}
\examples{
data(gene_coord)
}
\references{
Downloaded from
}
\keyword{datasets}
