% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mhc_genes}
\alias{mhc_genes}
\title{A Dataframe of genes and their genomic coordinates overlapping the MHC}
\format{
An object of class \code{"data.frame"};
}
\source{
Add later when paper published
}
\usage{
data(mhc_genes)
}
\description{
A data frame of genes extracted from ENSEMBL databses using REST API or BiomaRt.
}
\examples{
data(mhc_genes)
}
\references{
Add later when paper published
}
\keyword{datasets}
