% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_small}
\alias{seurat_small}
\title{A small scRNAseq matrix in Seurat format}
\format{
An object of class \code{"SeuratObject"}; see \code{\link[Seurat vingette]{https://satijalab.org/seurat/articles/get_started.html}}.
}
\source{
Add later when paper published
}
\usage{
data(seurat_small)
}
\description{
A Seurat object of 1452 cells and 23310 features (genes). It contains
cell type annotations at two levels of granularity that can be accessed
using seurat_small$cluster_level_1 and seurat_small$cluster_level_2.
}
\examples{
data(seurat_small)
}
\references{
Add later when paper published
}
\keyword{datasets}
