% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_genes.R
\name{get_mhc_genes}
\alias{get_mhc_genes}
\title{Get vector of hg38 genes overlapping MHC region using REST API or BiomaRt}
\usage{
get_mhc_genes(build = "hg38", method = "rest_api")
}
\arguments{
\item{build}{A genome build to query the MHC genes from (hg19 or hg38)}

\item{method}{The query method to use (biomart o rest_api)}
}
\value{
A dataframe of genes, positions and metadata overlapping the MHC region of chr6.
}
\description{
Get vector of hg38 genes overlapping MHC region using REST API or BiomaRt
}
