% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_genes.R
\name{get_mhc_genes}
\alias{get_mhc_genes}
\title{Get vector of MHC genes using biomaRt}
\usage{
get_mhc_genes()
}
\value{
A vector of unique genes that fall within MHC region of chr6.
}
\description{
Get vector of MHC genes using biomaRt
}
